feature(remove-env): Using Env not needed for react-native
Due to the complexity of adding a certificate on mobile, along with the
difficulties of using env files in react native (caching issues) we do
not need an env file for the react native portion. We also may not need
an env file at all if we go with a hosted server and have the user edit
the plist file directly.