task(feedback): Update Journey hook to remove extra access token call
In the ios and android bridge code, the login function will make the
call to get an access token, so it is not needed for us to make that
call seperately. We can rely on the bridge code to complete the oauth
flow and return us an access token and only call for an access token
when we need one.